Update of Habana SynapseAI notebook to 1.15.1
This required quite a few changes, since the SynapseAI stack
requires Python 3.11 for 1.15.1. Uses a similar workflow to
Anaconda with a custom base image, this however should be easy
to update in the future just by swapping out the FROM in the base
image and verifying the python versions going forward.

Signed-off-by: Sean Pryor <spryor@redhat.com>