[GHA] Enhance the checks for the kustomization.yaml files
Let's run on the currently used kustomize version by the operator but
also with the decently new kustomize version so we are sure we are
compatible with the upcoming features. This adds also STDERR output
check and compares the output between these two kustomize version.

Currently we have set kustomize 5.0.3 and 5.6.0.

Also this adds support for the execution on non-linux/non-amd64 archs.