Fixes for the digest updater for 2024b and 2024a images releases (#399)
* Fix the digest updater GHA regexp array initialization

After the removal of the habana image from the set of regexps in
red-hat-data-services/notebooks#384, there was kept a wrong syntax
causing this part of the GHA to fail. This change fixes this.

* Fix the regexps for the digest updater for 2024b images releases

The regexps updated was missed as part of the red-hat-data-services/notebooks#383
as there was nothing to test the GHA with before it got merged. This change
should fix it making the update for the 2024b working from now on.

* Update the notebook digest update GHA name and description