[tests] fix the pytest tests executio on linux machine with SELinux
There is necessary to use "z" option when mounting files into the
container otherwise SELinux will not allow container to read the mounted
file.

Also, I've adjusted the UIDs used in the tests so that they are lower
than 65k because otherwise it's not possible to run these with a
root-less podman/docker.