NO-JIRA: chore(tests/containers/workbenches/rstudio): trailing slash
Let's not use trailing slash for the directory. When this variable is
used and some value should be append to it, it should be done always as:
`APP_ROOT_HOME/some-path`. So the forward slash is added there and we
avoid having double slash in our outputs that may be misleading (e.g.
empty variable value in the path).