fix(gha): validate-dockerfiles more permissive
Presently, we use a `find . -name Dockerfile` to find files to lint with `Hadolint`.  With the move to remove chained builds - our Dockerfiles now have extensions that help differentiate what (if any) accelerator they are built for.

As a result, all of our Dockerfiles now have file extensions - which means the current code doesn't find them.

A simple `find . -name "Dockerfile*"` (adding a `*`) resolves this problem.

Note: Once we onboard into Konflux - we will also have `Dockerfile.konflux` files.  I have made the decision to **include** them in our linting (as they are part of our repo).  However, if that causes issue in the future - the following command could be used:
- `find . -name "Dockerfile*" -not -name "Dockerfile.konflux"`

Related-to: https://issues.redhat.com/browse/RHOAIENG-19048