RHOAIENG-19048: chore(gha): remove the upfront workflow generation script
With the chained build removal changes in, we need not worry about build chains in GHA.

* https://github.com/opendatahub-io/notebooks/pull/924

This means that there is no benefit from generating the image build workflow upfront and committing that.
Previously the benefit was that ad-hoc matrix jobs could not have dependencies on one another, whereas these generated jobs could.