NO-JIRA: chore(konflux): introduce image retention on pushed images
Until we decide to release on Konflux in upstream, it doesn't make much
sense to store the result images in the quay.io repository wasting the
disk resources there. Let's introduce the retention similarly as we do
for the PR builds. For PR we delete images after 5 days, let's have 28
days for the pushed builds just in case.

Once we decide to perform release in upstream from Konflux, this
decision needs to be revisited to avoid deletion of the released
artifacts.