NO-JIRA: fix(ci): update the way we check for the rocm version
With the introduction of the 2025.1 images, we changed the way we
install the ROCm on our images. Instead of installing the main `rocm`
metapackage, we install directly just it's subpackages ecluding the
installation of the `migraphx` and `mivisionx` packages to reduce the
size of the result ROCm images. This was a deliberate decision and so we
need to amend the way how to check the ROCm version on our images.

For the record:
* https://github.com/opendatahub-io/notebooks/pull/956/commits/0d262e27b599349f3f33e129f934584860ba4e24#diff-8ddc31b6fcd6307008d6abf6aef98768f83a7a94cb09a2eee2b3e1ff04788f20R40-R42