feat(kustomize): use LabelTransformer instead of deprecated commonLabels
This is a "piece" of a more comprehensive/interesting PR:
- https://github.com/opendatahub-io/notebooks/pull/1015

Unfortunately, that PR has grown wildly unwieldy in its size - and immediate feedback received was to try to break it into smaller pieces - so consider this one piece!

The ulitmate goal here on this targetted PR is two-fold:
- standardization irrespective of image build "flavour" our kustomize labelling
- get rid of following warning:

```
$ kubectl kustomize jupyter/minimal/ubi9-python-3.11/kustomize/base
...
```

No actual changes are introduced in this PR - simply leveraging the `LabelTransformer` to accomplish what `commonLabels` was previously doing.

Related-to: https://issues.redhat.com/browse/RHOAIENG-23291