NO-JIRA: tests(gha): fix `check-payload` to work with PRs and without `podman login` for root user (#1084)
Turns out, there were two problems with previous implementation, that were only revealed when

* the action was running on a push to a repository that has private github container registry
* the action was running on a pull request that modified some image

In the first case, action was unable to pull from repository that requires authentication.

In the second case, the image is not pushed anywhere, so attempts to pull are futile.

The solution appears to be what was implemented in

* https://github.com/openshift/check-payload/pull/154

That way we can run against a locally present image.