Allow install_pdf_deps.sh to work for multiarch
This will allow Dockerfiles that use this script to at least get the
correct texlive and pandoc for arm64 along with the existing amd64.

The corresponding change to Dockerfiles that use it here, are because
of the hardcoded x86_64 in the binary path: the script now renames
that directory to be more generic, so the references to it must all be
changed.