Add libxcrypt-compat to provide libcrypt.so.1
For this bunch of python 3.12 images it appears that
/opt/app-root/lib64/python3.12/site-packages/mysql/vendor/private/sasl2/libplain.so
exists and expects libcrypt.so.1 to also exist (if it is to be used at
least). I'm not sure why that doesn't exist in the equivalent 3.11 images.

It should be noted that libxcrypt-compat is deprecated in RHEL/UBI 9,
and how it's used should be considered before deciding to merge this
PR.