test(tests/containers): add test for ROCm accelerator images (#1414)
This commit introduces tests for ROCm-enabled workbench images on OpenShift. These tests verify that the images can be deployed successfully on a cluster with AMD GPUs and that both PyTorch and TensorFlow can correctly detect the available accelerator.
To support the testing of large accelerator images, the following changes were made:
- The pod readiness timeout in the test framework has been increased to 10 minutes to allow sufficient time for image pulling.
- The utility was updated to allow for configurable timeouts. `ImageDeployment`
- Existing CUDA tests were updated to use this new configurable timeout.