[2024b] RHOAIENG-28774: update the 2024b GitHub Action so that it does not end up pulling the wrong cache
This is a selective cherry-pick from the

* https://github.com/opendatahub-io/notebooks/pull/1330

because otherwise the x86-64 build may end up fetching arm64 cache and things then would not work.

First, the arm64 cache is too big to fit (2024b partitions disk differently, less space on /)
and second, even if it did unpack the cache, it would not be able to run arm64 binaries.