NO-JIRA: ensure empty lines are skipped while processing env files in `check-params-env.sh` (#1784)
This solves the issues in https://github.com/opendatahub-io/notebooks/actions/runs/16837438136/job/47700268773

triggered by introduction of empty lines in `params-latest.env` in

* https://github.com/opendatahub-io/notebooks/pull/1730

```
Checking that all variables in the file 'manifests/base/params.env' & 'manifests/base/params-latest.env' are unique and expected
Some of the variables in the file aren't unique!
```

```
Checking that all values assigned to variables in the file 'manifests/base/params.env' & 'manifests/base/params-latest.env' are unique and expected
Some of the values in the file aren't unique!
```

```
Checking that there are expected number of records in the file 'manifests/base/params.env' + 'manifests/base/params-latest.env'
Number of records in the file is incorrect - expected '44' but got '75'!
```

```
Checking format of: ''
ERROR: Line doesn't contain '=' and it should!
```