Allow pytorch images to build on aarch64
This changes the pytorch images to use the CPU torchvision wheel on
aarch64, since pytorch doesn't provide an aarch64 cuda build for that
version (or any version?) of torchvision.

For the pytorch+llmcompressor images, this changes both torch and
torchvision to use CPU wheels, because pytorch doesn't provide an
aarch cuda version even for torch at this version.

This changes also uses the `--universal` flag to the `uv pip compile`
command to create the pylock.toml file. This is so that no matter
which platform this command is run from, it will include the specific
versions of these dependencies for the corresponding target platform.

Related to this, this change also explicitly adds the x86-64-only
nvidia wheels explicitly to the pyproject.toml. This seems to be
necessary because the x86_64 wheel specifies them as `Requires-Dist`
in the METADATA file in the wheel, but the aarch64 wheel does
not. This seems like a bug in those wheels, as in later wheel
versions (e.g. torch-2.8.0+cu129-cp312-cp312-manylinux_2_28-*)
those dependencies are specified in the wheels for both arches, but
both have markers to specify that they're only actually expected to be
installed on x86_64.

The `uv pip check` after the `uv pip install` in the Dockerfile is a
quick sanity check to make sure that the right dep set is
installed. This should help prevent bad locking (it's one way to catch
the missing nvidia deps on x86_64).