RHOAIENG-24702, RHOAIENG-38121: add AI Pipelines exceptions to check-payload (#2654)
These exceptions are required because AI Pipelines uses two different
binaries when in FIPS mode vs non-FIPS mode. The reason is that these
binaries are copied to the user's provided base image and in FIPS mode,
this means the OpenSSL version must match (e.g. UBI 9). In non-FIPS
mode, we leverage the builtin Go crypto so that the binaries are
portable.

Relates:
RHOAIENG-24702 and RHOAIENG-38121

Slack:
https://redhat-internal.slack.com/archives/C09GK0M9YKD/p1762370844652579

Signed-off-by: mprahl <mprahl@users.noreply.github.com>