RHAIENG-2042: chore(base-images): add `TARGETARCH` as an argument to `ubi9-python-3.12` Dockerfile
The c9s PR did not trigger building the ubi9 version, so I missed on the fact that changes are needed there.

It seems reasonable to remove the ubi9 base image as it is unused and only leave c9s. But I don't want to do it right now, first I will make it work and then make it go away.